backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.12.1
python-multipart==0.0.6
groq==0.4.2
# SQLite support (included with Python) 